use _shared_docs
